moved RepositoryInfoDto
