+++ src/main/java/org/project/githubrepositories/repository/RepositoryInfoController.java
@@ -1,6 +1,7 @@
 package org.project.githubrepositories.repository;
 
 import lombok.AllArgsConstructor;
+import org.project.githubrepositories.repository.dto.RepositoryInfoDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
